api/requirements.txt
amadeus
flask-restful
requests
